Split bump integration tests into cohesive modules
Resolve the CodeScene low-cohesion finding (16 responsibilities across 23
functions) for tests/unit/test_bump_command_integration.py by splitting it
into four focused modules:

- test_bump_manifest_updates.py     - manifest version/dependency updates
- test_bump_lockfile_rebuild.py     - lockfile rebuild behaviour
- test_bump_rebuild_lockfiles_resolution.py - issue #106 resolution matrix
- test_bump_documentation_updates.py - README/doc transposition

Promote the shared stub_lockfile_regeneration fixture to tests/unit/conftest.py
as an autouse fixture. It is scoped by module name to the four bump-run split
modules so it never shadows test_bump_lockfiles, which exercises
regenerate_lockfiles directly and needs the real implementation. Test logic,
assertions, parametrisation, and the fixture body are unchanged; the full unit
suite still reports 569 passed.
